# rudi-microservice/rudi-microservice-acl/openapi/rudi-acl-api.json
{
	"openapi": "3.0.0",
	"info": {
		"description": "API Public ACL Rudi",
		"version": "0.0.1",
		"title": "rudi-acl-api",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"email": "contact@rudi.fr"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"servers": [
		{
			"url": "/{basePath}",
			"description": "Rudi Acl API",
			"variables": {
				"basePath": {
					"default": "acl/v1"
				}
			}
		}
	],
...
	"paths": {
...
		"/oauth2-authenticators": {
			"get": {
				"tags": [
					"authenticator"
				],
				"summary": "Retoure la configuration des authenticators",
				"description": "Retoure la configuration des authenticators",
				"operationId": "getOAuth2Authenticators",
				"parameters": [],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "rudi-acl-model.json#/components/schemas/OAuth2AuthenticatorDescription"
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/oauth2-authenticators/icons/{authenticator-name}": {
			"get": {
				"tags": [
					"authenticator"
				],
				"summary": "Retoure l'icone associé à l'authenticator",
				"description": "Retoure l'icone associé à l'authenticator",
				"operationId": "getOAuth2AuthenticatorIcon",
				"parameters": [
					{
						"in": "path",
						"name": "authenticator-name",
						"description": "La clé de l'authenticator",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/octet-stream": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/oauth2-authenticators/{authenticator-name}": {
			"get": {
				"tags": [
					"authenticator"
				],
				"summary": "Retoure la configuration des authenticators",
				"description": "Retoure la configuration des authenticators",
				"operationId": "authenticateThroughAuthenticator",
				"parameters": [
					{
						"in": "path",
						"name": "authenticator-name",
						"description": "La clé de l'authenticator",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
...
		"/user-info-uuid/{userUuid}": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Retourne un utilisateur avec les données minimales pour un affichage web",
				"description": "Retourne un utilisateur avec les données minimales pour un affichage web",
				"operationId": "getUserInfo",
				"parameters": [
					{
						"name": "userUuid",
						"in": "path",
						"description": "uuid pour la récupération d'un élément",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/User"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/user-info-login/{userLogin}": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Retourne un utilisateur avec les données minimales pour un affichage web",
				"description": "Retourne un utilisateur avec les données minimales pour un affichage web",
				"operationId": "getUserInfoByLogin",
				"parameters": [
					{
						"name": "userLogin",
						"in": "path",
						"required": true,
						"description": "login pour la récupération d'un élément",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/User"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
...
		"/roles": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Collection de roles d'utilisateurs",
				"description": "Retourne une liste de roles d'utilisateurs",
				"operationId": "searchRoles",
				"parameters": [
					{
						"in": "query",
						"name": "code",
						"description": "Tout ou partie du code du rôle",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "label",
						"description": "Tout ou partie du label du rôle",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "active",
						"description": "Elements actif ou tous",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "rudi-acl-model.json#/components/schemas/Role"
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"acl"
				],
				"summary": "Ajout d'un nouveau role d'utilisateurs",
				"description": "Permet d'ajouter un nouveau role d'utilisateurs",
				"operationId": "createRole",
				"requestBody": {
					"description": "info role",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "rudi-acl-model.json#/components/schemas/Role"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/Role"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": [
					"acl"
				],
				"summary": "Mise à jour d'un role d'utilisateurs",
				"description": "Mise à jour d'un role d'utilisateurs",
				"operationId": "updateRole",
				"requestBody": {
					"description": "info role",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "rudi-acl-model.json#/components/schemas/Role"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/Role"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/roles/{uuid}": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Retourne un role d'utilisateurs",
				"description": "Retourne un role d'utilisateurs",
				"operationId": "getRole",
				"parameters": [
					{
						"name": "uuid",
						"in": "path",
						"description": "uuid role",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/Role"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": [
					"acl"
				],
				"summary": "Suppression d'un role d'utilisateurs",
				"description": "Suppression d'un role d'utilisateurs",
				"operationId": "deleteRole",
				"parameters": [
					{
						"name": "uuid",
						"in": "path",
						"description": "uuid role",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"ApiError": {
				"type": "object",
				"properties": {
					"code": {
						"type": "string"
					},
					"label": {
						"type": "string"
					}
				}
			},
...
		}
	}
}